move as const to context and reduce prop dirlling
